숙소 찜하기 기능 관련 코드 모두 Accommodation~ 클래스, xml로 이동
