+++ src/main/java/kr/co/nc/mapper/AccommodationMapper.java
@@ -1,6 +1,7 @@
 package kr.co.nc.mapper;
 
 import java.util.List;
+import java.util.Map;
 
 import org.apache.ibatis.annotations.Mapper;
 
@@ -50,4 +51,12 @@ public interface AccommodationMapper {
 	List<String> getImagesByRoomNo(int roomNo);
 	// 검색 기간에 따른 특정 객실의 예약 가능 여부 조회 (가능하면 1, 불가능하면 0을 반환)
 	Integer getAvailableRoomStock(RoomCriteria criteria);
+	
+	
+	// 해당 사용자의, 해당 번호의 숙소에 대한 찜하기 정보가 존재하면 1을, 존재하지 않으면 0을 반환
+	int isExistUserLikeByAccoId(Map<String, Integer> param);
+	void insertUserLikeByAccoId(Map<String, Integer> param);
+	void deleteUserLikeByAccoId(Map<String, Integer> param);
+	// 사용자가 찜하기 누른 모든 숙소 정보 반환
+	List<Accommodation> getAllLikedAccoByUserNo(int no); 
 }
