+++ src/main/java/kr/co/nc/mapper/UserMapper.java
@@ -1,7 +1,5 @@
 package kr.co.nc.mapper;
 
-import java.util.Map;
-
 import org.apache.ibatis.annotations.Mapper;
 
 import kr.co.nc.vo.User;
@@ -14,10 +12,5 @@ public interface UserMapper {
 	User getUserById(String id);
 	User getUserByEmail(String email);
 	void update(User user);
-	
-	// 해당 찜하기 정보가 존재하면 1을, 존재하지 않으면 0을 반환
-	int isExistUserLikeByAccoId(Map<String, Integer> param);
-	void insertUserLikeByAccoId(Map<String, Integer> param);
-	void deleteUserLikeByAccoId(Map<String, Integer> param);
 	
 }
