+++ src/main/resources/mybatis/mappers/AccommodationMapper.xml
@@ -185,6 +185,7 @@
 		where acco_id = #{value}
 	</select>
 	
+	<!-- 객실 정보 관련 -->
 	<!-- RoomResultMap -->
 	<resultMap type="AccommodationRoom" id="RoomResultMap">
 		<id column="room_no" property="no" />
@@ -229,4 +230,26 @@
 		where fr.room_no = nr.room_no (+)
 		and fr.room_no = #{roomNo}
 	</select>
+	
+	<!-- 찜하기 기능 관련 -->
+	<!-- int isExistUserLikeByAccoId(Map<String, Integer> param); -->
+	<select id="isExistUserLikeByAccoId" parameterType="map" resultType="int">
+		select count(*)
+		from user_accommodation_likes
+		where user_no = #{userNo}
+			and acco_id = #{accoId}
+	</select>
+	
+	<!-- void insertUserLikeByAccoId(Map<String, Integer> param); -->
+	<insert id="insertUserLikeByAccoId" parameterType="map">
+		insert into user_accommodation_likes (user_no, acco_id)
+		values (#{userNo}, #{accoId})
+	</insert>
+	
+	<!-- void deleteUserLikeByAccoId(Map<String, Integer> param); -->
+	<delete id="deleteUserLikeByAccoId" parameterType="map">
+		delete from user_accommodation_likes
+		where user_no = #{userNo}
+			and acco_id = #{accoId}
+	</delete>
 </mapper>
